Configure compile task mixin options lazily (#1256)
* Try to configure compile task mixin options lazily

* Fix CompileJava being realised too soon, and other misc lazy task fixes.
